Tests/LibWeb/Text/expected/css/element-reference-pseudo-stale.txt

before change: rgb(0, 128, 0)
after change: rgb(128, 128, 128)
